Algorithm improvements and minor changes.
Now login and logout functions load logFile in RAM in a dynamic string array. No more use of a temporany file. This will avoid risk of tragedies such as `people walking on the logFile`. It was also implemented a function `createBackupIfNecessary` that creates a backup copy of the logFile once a month.